Moved `Module` to its own module
